Error in complain has been resolved and going on at get friends charities end point

+++ src/main/java/kg/giftlist/giftlistm2/db/service/WishListService.java
@@ -2,10 +2,7 @@ package kg.giftlist.giftlistm2.db.service;
 
 import kg.giftlist.giftlistm2.controller.payload.WishListRequest;
 import kg.giftlist.giftlistm2.controller.payload.WishListResponse;
-import kg.giftlist.giftlistm2.db.entity.Booking;
-import kg.giftlist.giftlistm2.db.entity.Holiday;
-import kg.giftlist.giftlistm2.db.entity.User;
-import kg.giftlist.giftlistm2.db.entity.WishList;
+import kg.giftlist.giftlistm2.db.entity.*;
 import kg.giftlist.giftlistm2.db.repository.*;
 import kg.giftlist.giftlistm2.enums.WishListStatus;
 import kg.giftlist.giftlistm2.exception.BadCredentialsException;
@@ -44,6 +41,7 @@ public class WishListService {
         } else {
             wishList.setGiftName(request.getGiftName());
         }
+        wishList.setBlocked(false);
         wishList.setUser(user);
         wishList.setLink(request.getLink());
         wishList.setImage(request.getImage());
@@ -52,8 +50,8 @@ public class WishListService {
         wishList.setDescription(request.getDescription());
         wishList.setWishListStatus(WishListStatus.NOT_BOOKED);
         wishListRepository.save(wishList);
-        List<User>friendList = userRepository.getAllFriendByUserId(user.getId());
-        wishList.addNotification(notificationService.wishListNotification(user,friendList,wishList));
+        List<User> friendList = userRepository.getAllFriendByUserId(user.getId());
+        wishList.addNotification(notificationService.wishListNotification(user, friendList, wishList));
         notificationRepository.saveAll(wishList.getNotifications());
         return mapToResponse(wishList);
     }
@@ -65,43 +63,52 @@ public class WishListService {
         if (user.getWishLists().contains(wishList)) {
             throw new WishListExistException("You have this wishList");
         }
-        WishListRequest wishListRequest = new WishListRequest();
-        wishListRequest.setGiftName(wishList.getGiftName());
-        wishListRequest.setDescription(wishList.getDescription());
-        wishListRequest.setLink(wishList.getLink());
-        wishListRequest.setHolidayDate(wishListRequest.getHolidayDate());
-        wishListRequest.setImage(wishList.getImage());
-        wishListRequest.setHolidayId(wishList.getHolidays().getId());
-        return create(wishListRequest);
+        if (wishList.isBlocked()) {
+            throw new BadCredentialsException("This wish list was blocked due to it got a complain. Contact to administration of Giftlist");
+        } else {
+            WishListRequest wishListRequest = new WishListRequest();
+            wishListRequest.setGiftName(wishList.getGiftName());
+            wishListRequest.setDescription(wishList.getDescription());
+            wishListRequest.setLink(wishList.getLink());
+            wishListRequest.setHolidayDate(wishListRequest.getHolidayDate());
+            wishListRequest.setImage(wishList.getImage());
+            wishListRequest.setHolidayId(wishList.getHolidays().getId());
+            return create(wishListRequest);
+        }
     }
 
     public WishListResponse update(Long id, WishListRequest request) {
         User user = getAuthenticatedUser();
         if (wishListRepository.findById(id).isEmpty()) {
             throw new EmptyValueException("There is no any wish list with id " + id);
-        }
-        WishList wishList = wishListRepository.findById(id).get();
-        if (user.getWishLists().contains(wishList)) {
-            Holiday holiday = holidayRepository.findById(request.getHolidayId()).get();
-            if (request.getGiftName().isEmpty()) {
-                throw new EmptyValueException("Wish list name must not be empty!");
-            }
-            wishList.setGiftName(request.getGiftName());
-            wishList.setLink(request.getLink());
-            wishList.setImage(request.getImage());
-            wishList.setHolidays(holiday);
-            wishList.setHolidayDate(request.getHolidayDate());
-            wishList.setDescription(request.getDescription());
-            if (bookingRepository.findById(wishList.getId()).isPresent()) {
-                wishList.setWishListStatus(WishListStatus.BOOKED);
-            }
-            if (bookingRepository.findById(wishList.getId()).isEmpty()) {
-                wishList.setWishListStatus(WishListStatus.NOT_BOOKED);
-            }
-            wishListRepository.save(wishList);
-            return mapToResponse(wishList);
         } else {
-            throw new EmptyValueException("You have no any wish list with id " + id);
+            WishList wishList = wishListRepository.findById(id).get();
+            if (wishList.isBlocked()) {
+                throw new BadCredentialsException("This wish list was blocked due to it got a complain. Contact to administration of Giftlist");
+            } else {
+                if (user.getWishLists().contains(wishList)) {
+                    Holiday holiday = holidayRepository.findById(request.getHolidayId()).get();
+                    if (request.getGiftName().isEmpty()) {
+                        throw new EmptyValueException("Wish list name must not be empty!");
+                    }
+                    wishList.setGiftName(request.getGiftName());
+                    wishList.setLink(request.getLink());
+                    wishList.setImage(request.getImage());
+                    wishList.setHolidays(holiday);
+                    wishList.setHolidayDate(request.getHolidayDate());
+                    wishList.setDescription(request.getDescription());
+                    if (bookingRepository.findById(wishList.getId()).isPresent()) {
+                        wishList.setWishListStatus(WishListStatus.BOOKED);
+                    }
+                    if (bookingRepository.findById(wishList.getId()).isEmpty()) {
+                        wishList.setWishListStatus(WishListStatus.NOT_BOOKED);
+                    }
+                    wishListRepository.save(wishList);
+                    return mapToResponse(wishList);
+                } else {
+                    throw new EmptyValueException("You have no any wish list with id " + id);
+                }
+            }
         }
     }
 
@@ -109,19 +116,24 @@ public class WishListService {
         User user = getAuthenticatedUser();
         if (wishListRepository.findById(id).isEmpty()) {
             throw new EmptyValueException("There is no any wish list with id " + id);
-        }
-        WishList wishList = wishListRepository.findById(id).get();
-        if (user.getWishLists().isEmpty()) {
-            throw new EmptyValueException("You have no any wish list with id " + id);
-        }
-        if (user.getWishLists().contains(wishList)) {
-            log.info("wishList" + wishList.getGiftName());
-            user.getWishLists().remove(wishList);
-            wishListRepository.delete(wishList);
         } else {
-            throw new EmptyValueException("You have no any wish list with id " + id);
+            WishList wishList = wishListRepository.findById(id).get();
+            if (wishList.isBlocked()) {
+                throw new BadCredentialsException("This wish list was blocked due to it got a complain. Contact to administration of Giftlist");
+            } else {
+                if (user.getWishLists().isEmpty()) {
+                    throw new EmptyValueException("You have no any wish list with id " + id);
+                }
+                if (user.getWishLists().contains(wishList)) {
+                    log.info("wishList" + wishList.getGiftName());
+                    user.getWishLists().remove(wishList);
+                    wishListRepository.delete(wishList);
+                    return "Wish list successfully was deleted!";
+                } else {
+                    throw new EmptyValueException("You have no any wish list with id " + id);
+                }
+            }
         }
-        return "Wish list successfully was deleted!";
     }
 
     public WishListResponse getWishListById(Long id) {
@@ -146,39 +158,51 @@ public class WishListService {
         return view(wishLists);
     }
 
-    public String book(Long id) {
-        User user = getAuthenticatedUser();
-        if (wishListRepository.findById(id).isEmpty()) {
-            throw new EmptyValueException("There is no wishList with id " + id);
-        }
-        WishList wishList = wishListRepository.findById(id).get();
-        if (wishList.getWishListStatus().equals(WishListStatus.NOT_BOOKED)) {
-            Booking booking1 = new Booking();
-            booking1.setId(booking1.getId());
-            booking1.setWishList(wishList);
-            booking1.setUserId(user);
-            bookingRepository.save(booking1);
-            wishList.setWishListStatus(WishListStatus.BOOKED);
-            wishListRepository.save(wishList);
-            return "You have successfully booked this wishList";
+    public WishListResponse getWishlistByAdmin(Long id) {
+        if (wishListRepository.existsById(id)) {
+            WishList wishList = wishListRepository.findById(id).get();
+            return mapToResponse(wishList);
         } else {
-            throw new BadCredentialsException("This wishList is already booked");
+            throw new EmptyValueException("There is no any wish list with id " + id);
         }
     }
 
-    public String unBook(Long id) {
-        User user = getAuthenticatedUser();
-        if (bookingRepository.findById(id).isEmpty()) {
-            throw new EmptyValueException("There is no wish list with id " + id);
-        }
-        Booking booking = bookingRepository.findById(id).get();
-        if (user.getBookings().contains(booking)) {
-            user.getBookings().remove(booking);
-            bookingRepository.delete(booking);
-            booking.getWishList().setWishListStatus(WishListStatus.NOT_BOOKED);
-            return "You have successfully unbooked this wish list";
+    public List<WishListResponse> getAllWishlistsByAdmin() {
+        if (wishListRepository.findAll().isEmpty()) {
+            throw new EmptyValueException("There is no any wish list");
         } else {
-            throw new BadCredentialsException("This wish list is already unbooked");
+            List<WishList> wishLists = wishListRepository.findAll();
+            return view(wishLists);
+        }
+    }
+
+    public String blockWishlistByAdmin(Long id) {
+        if (wishListRepository.findById(id).isEmpty()) {
+            throw new EmptyValueException("There is no any wish list with id " + id);
+        } else {
+            WishList wishList = wishListRepository.findById(id).get();
+            if (wishList.isBlocked()) {
+                throw new BadCredentialsException("You have already blocked the wish list with id " + id);
+            } else {
+                wishList.setBlocked(true);
+                wishListRepository.save(wishList);
+                return "You have blocked the wish list with id " + id;
+            }
+        }
+    }
+
+    public String unBlockWishlistByAdmin(Long id) {
+        if (wishListRepository.findById(id).isEmpty()) {
+            throw new EmptyValueException("There is no any wish list with id " + id);
+        } else {
+            WishList wishList = wishListRepository.findById(id).get();
+            if (!wishList.isBlocked()) {
+                throw new BadCredentialsException("You have already unblocked the wish list with id " + id);
+            } else {
+                wishList.setBlocked(false);
+                wishListRepository.save(wishList);
+                return "You have unblocked the wish list with id " + id;
+            }
         }
     }
